feat: move PricingConfig to aicore/pricing.py and update imports
